pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stock-research-tool"
version = "0.1.0"
description = "AI-Powered Stock Research Tool with fund tracking and analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Stock Research Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # FastAPI and web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",

    # Redis and caching
    "redis>=5.0.1",
    "hiredis>=2.2.3",

    # Task queue
    "celery[redis]>=5.3.4",

    # Configuration and validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",

    # HTTP clients
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",

    # Financial data sources
    "yfinance>=0.2.33",
    "alpha-vantage>=2.3.1",

    # AI/LLM
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-community>=0.0.10",
    "ollama>=0.1.3",

    # Data processing
    "pandas>=2.1.3",
    "numpy>=1.26.2",

    # PDF generation
    "weasyprint>=60.1",
    "jinja2>=3.1.2",

    # Charting
    "plotly>=5.18.0",
    "kaleido>=0.2.1",

    # Retry and resilience
    "tenacity>=8.2.3",

    # Logging and monitoring
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Utilities
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "bandit>=1.7.6",
    "safety>=2.3.5",
]

dagster = [
    "dagster>=1.5.10",
    "dagster-webserver>=1.5.10",
    "dagster-postgres>=0.21.10",
]

[project.scripts]
stocktool = "backend.app.main:run"

[tool.hatch.build.targets.wheel]
packages = ["backend", "agents", "pipelines"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests", "agents/tests", "pipelines/tests"]
addopts = "-v --cov=backend --cov=agents --cov=pipelines --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # defaults
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
exclude = ["migrations", "tests"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101"]  # assert_used

[tool.ruff.isort]
known-first-party = ["backend", "agents", "pipelines"]

[tool.coverage.run]
source = ["backend", "agents", "pipelines"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
